Renamed modules back to modal.*.
