+++ packages/core/src/lib/modal/modal.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import ReactDOM from "react-dom";
 
 import { WalletSelector } from "../wallet-selector.types";
-import { WalletSelectorModal, ModalOptions } from "./setupModal.types";
+import { WalletSelectorModal, ModalOptions } from "./modal.types";
 import { Modal } from "./components/Modal";
 
 const MODAL_ELEMENT_ID = "near-wallet-selector-modal";
